Rename data model module for consistency with instrument_model
