+++ tests/test_hdf_writer.py
@@ -2,7 +2,7 @@ import h5py
 from math import sqrt
 from pytest import approx
 from geometry_constructor.writers import HdfWriter
-from geometry_constructor.datamodel import Detector, PixelGrid, Corner, CountDirection, Vector
+from geometry_constructor.data_model import Detector, PixelGrid, Corner, CountDirection, Vector
 from geometry_constructor.instrument_model import InstrumentModel
