+++ tests/test_instrument_model.py
@@ -1,18 +1,18 @@
-from geometry_constructor import datamodel
+from geometry_constructor import data_model
 from geometry_constructor.instrument_model import InstrumentModel
 
 
 def test_initialise_model():
     model = InstrumentModel()
     assert model.rowCount() == 1
-    assert isinstance(model.components[0], datamodel.Sample)
+    assert isinstance(model.components[0], data_model.Sample)
 
 
 def test_add_component():
     model = InstrumentModel()
     model.add_detector('My Detector')
     assert model.rowCount() == 2
-    assert isinstance(model.components[1], datamodel.Detector)
+    assert isinstance(model.components[1], data_model.Detector)
     assert model.components[1].name == 'My Detector'
 
 
@@ -21,4 +21,4 @@ def test_remove_component():
     model.add_detector('My Detector')
     model.remove_component(1)
     assert model.rowCount() == 1
-    assert not isinstance(model.components[0], datamodel.Detector)
+    assert not isinstance(model.components[0], data_model.Detector)
